+++ previous-versions/sync-for-expenses-version-1/RELEASES.md
@@ -19,3 +19,13 @@ Based on:
 - [typescript v0.1.1] previous-versions/sync-for-expenses-version-1
 ### Releases
 - [NPM v0.1.1] https://www.npmjs.com/package/@codat/sync-for-expenses-version-1/v/0.1.1 - previous-versions/sync-for-expenses-version-1
+
+## 2023-09-05 00:14:47
+### Changes
+Based on:
+- OpenAPI Doc prealpha https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Sync-Expenses-v1.yaml
+- Speakeasy CLI 1.78.3 (2.96.3) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.2.0] previous-versions/sync-for-expenses-version-1
+### Releases
+- [NPM v0.2.0] https://www.npmjs.com/package/@codat/sync-for-expenses-version-1/v/0.2.0 - previous-versions/sync-for-expenses-version-1